+++ CHANGELOG.md
@@ -25,6 +25,10 @@
   ([PR #937](https://github.com/jasmin-lang/jasmin/pull/937);
   fixes [#895](https://github.com/jasmin-lang/jasmin/issues/895)).
 
+- Improve error message when renaming is not possible due to liveness conflicts
+  ([PR #940](https://github.com/jasmin-lang/jasmin/pull/940);
+  fixes [#483](https://github.com/jasmin-lang/jasmin/issues/483)).
+
 ## Other changes
 
 - The deprecated legacy interface to the LATEX pretty-printer has been removed